requirements.txt
paho-mqtt
requests
h5py
pytz